Export name field into common model
