--- src/aap_eda/core/models/base_org.py
@@ -1,31 +1,0 @@
-#  Copyright 2024 Red Hat, Inc.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from django.db import models
-
-from .utils import get_default_organization_id
-
-__all__ = ("BaseOrgModel",)
-
-
-class BaseOrgModel(models.Model):
-    class Meta:
-        abstract = True
-
-    organization = models.ForeignKey(
-        "Organization",
-        on_delete=models.CASCADE,
-        default=get_default_organization_id,
-        null=True,
-    )
