+++ src/aap_eda/core/models/credential_type.py
@@ -14,13 +14,12 @@
 
 from django.db import models
 
-from .base import BaseModel
-from .base_org import BaseOrgModel
+from .base import BaseOrgModel, UniqueNamedModel
 
 __all__ = ("CredentialType",)
 
 
-class CredentialType(BaseOrgModel, BaseModel):
+class CredentialType(BaseOrgModel, UniqueNamedModel):
     class Meta:
         db_table = "core_credential_type"
         constraints = [
@@ -30,7 +29,6 @@ class CredentialType(BaseOrgModel, BaseModel):
             )
         ]
 
-    name = models.TextField(null=False, unique=True)
     description = models.TextField(default="", blank=True, null=False)
     inputs = models.JSONField(default=dict)
     injectors = models.JSONField(default=dict)
