+++ src/aap_eda/core/models/decision_environment.py
@@ -14,13 +14,12 @@
 
 from django.db import models
 
-from .base import BaseModel
-from .base_org import BaseOrgModel
+from .base import BaseOrgModel, UniqueNamedModel
 
 __all__ = ("DecisionEnvironment",)
 
 
-class DecisionEnvironment(BaseOrgModel, BaseModel):
+class DecisionEnvironment(BaseOrgModel, UniqueNamedModel):
     class Meta:
         db_table = "core_decision_environment"
         constraints = [
@@ -30,7 +29,6 @@ class DecisionEnvironment(BaseOrgModel, BaseModel):
             )
         ]
 
-    name = models.TextField(null=False, unique=True)
     description = models.TextField(default="", blank=True, null=False)
     image_url = models.TextField(blank=False, null=False)
     # TODO: used by migration, remove it later
