+++ src/aap_eda/core/models/project.py
@@ -28,13 +28,12 @@ from django.db import models
 
 from aap_eda.core.utils.crypto.fields import EncryptedTextField
 
-from .base import BaseModel
-from .base_org import BaseOrgModel
+from .base import BaseOrgModel, UniqueNamedModel
 
 PROJECT_ARCHIVE_DIR = "projects/"
 
 
-class Project(BaseOrgModel, BaseModel):
+class Project(BaseOrgModel, UniqueNamedModel):
     class Meta:
         db_table = "core_project"
         constraints = [
@@ -53,11 +52,6 @@ class Project(BaseOrgModel, BaseModel):
     class ScmType(models.TextChoices):
         GIT = "git"
 
-    name = models.TextField(
-        null=False,
-        unique=True,
-        error_messages={"unique": "A project with this name already exists."},
-    )
     description = models.TextField(default="", blank=True, null=False)
     url = models.TextField(null=False)
     proxy = EncryptedTextField(blank=True, default="")
