+++ src/aap_eda/core/models/rulebook.py
@@ -15,8 +15,7 @@
 import yaml
 from django.db import models
 
-from .base import BaseModel
-from .base_org import BaseOrgModel
+from .base import BaseOrgModel
 
 __all__ = (
     "Rulebook",
@@ -26,7 +25,7 @@ __all__ = (
 )
 
 
-class Rulebook(BaseOrgModel, BaseModel):
+class Rulebook(BaseOrgModel):
     class Meta:
         db_table = "core_rulebook"
         unique_together = ["project_id", "name"]
@@ -87,7 +86,7 @@ class Rule(models.Model):
     action = models.JSONField(default=dict, null=False)
 
 
-class AuditRule(BaseOrgModel, BaseModel):
+class AuditRule(BaseOrgModel):
     class Meta:
         db_table = "core_audit_rule"
         indexes = [
